requirements.txt
pandas
numpy
matplotlib
seaborn
yfinance
requests
python-dotenv
sec-edgar-downloader
beautifulsoup4
scikit-learn
statsmodels
nltk
